fixing all the things
