+++ lib/src/types/shaders.d.ts
@@ -1,0 +1,5 @@
+export declare class Shader {
+    id: number;
+    locs: number[];
+    constructor(id: number, locs: number[]);
+}
