+++ lib/src/types/shaders.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var Shader = /** @class */ (function () {
+    function Shader(id, locs) {
+        this.id = id;
+        this.locs = locs;
+    }
+    return Shader;
+}());
+exports.Shader = Shader;
