+++ src/types/shaders.ts
@@ -1,0 +1,9 @@
+export class Shader {
+  id: number
+  locs: number[]
+
+  constructor(id: number, locs: number[]) {
+    this.id = id
+    this.locs = locs
+  }
+}
